+++ src/app/app.module.ts
@@ -4,9 +4,9 @@ import { StoreModule, combineReducers} from '@ngrx/store';
 import { EffectsModule} from '@ngrx/effects';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { KeepstraightApp } from './app.component';
-import { ScoreTable } from '../pages/game/scoretable.component';
-import { SetupControls } from '../pages/setup/gamesetup.component';
-import { UpdateScore } from '../pages/game/updatescore.modal';
+import { ScoreTable } from '../pages/game/components/scoretable.component';
+import { SetupControls } from '../pages/setup/components/gamesetup.component';
+import { UpdateScore } from '../pages/game/components/updatescore.modal';
 import { SetupPage } from '../pages/setup/setup';
 import { GamePage } from '../pages/game/game';
 import { SetupReducer } from './reducers/setup-reducer';
