+++ src/pages/game/components/scoretable.component.html
