+++ src/pages/game/components/scoretable.component.scss
