+++ src/pages/game/components/scoretable.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input } from '@angular/core';
 import { Observable } from 'rxjs/rx';
-import {AppState} from "../../app/services/app-state";
+import {AppState} from "../../../app/services/app-state";
 import {Store} from "@ngrx/store";
-import {GameActions} from "../../app/actions/game-actions";
+import {GameActions} from "../../../app/actions/game-actions";
 
 @Component({
   selector: 'score-table',
