+++ src/pages/setup/components/gamesetup.component.html
